+++ .env.example
@@ -21,7 +21,7 @@ VOICEFLOW_ASSEMBLYAI_API_KEY='your_assemblyai_api_key' # AssemblyAI API 密钥
 # 语音服务端口配置
 VOICEFLOW_SERVER_PORT=18080                    # VoiceFlow 服务端口
 
-# Huoshan 配置
-VOICEFLOW_HUOSHAN_ACCESS_KEY=''
-VOICEFLOW_HUOSHAN_APP_KEY=''
-VOICEFLOW_HUOSHAN_WS_URL='wss://openspeech.bytedance.com/api/v3/sauc/bigmode
+# VOLCENGINE 配置
+VOICEFLOW_VOLCENGINE_ACCESS_KEY=''
+VOICEFLOW_VOLCENGINE_APP_KEY=''
+VOICEFLOW_VOLCENGINE_WS_URL='wss://openspeech.bytedance.com/api/v3/sauc/bigmode
